Reorganize SqlBulkHelpers into their own class library for better code organization.  Add new Unit Test project for development of unit tests.
FIXED significant issue with where SqlBulkCopy.WriteToServer() method from SqlClient package does not guarantee the order of data being written to the server -- and there is still no support for the ORDER() hint -- therefore the data being written and the Identity values being assigned would not correctly correlate as expected for the list of data being sent.  Generally this was resulting in the WriteToServer() method inverting the order of data being sent in Bulk which then resulted in Identity values being incorrect based on the order of data specified. For example: if you sent 3 items for rows 1,2,3 then the identity values resulting from a new Insert would be out of order such as 21,20,19 instead of the expected 19,20,21.  Therefore when querying the data and attempting to sort by the Identity column to preserve the original order the results would be incorrect.